+++ glide-core/redis-rs/redis/src/cluster_async/mod.rs
@@ -31,7 +31,7 @@ pub mod testing {
 }
 use crate::{
     client::GlideConnectionOptions,
-    cluster_routing::{Routable, RoutingInfo},
+    cluster_routing::{Routable, RoutingInfo, ShardUpdateResult},
     cluster_slotmap::SlotMap,
     cluster_topology::{
         calculate_topology, get_slot, SlotRefreshState, DEFAULT_NUMBER_OF_REFRESH_SLOTS_RETRIES,
@@ -55,6 +55,7 @@ use std::{
     task::{self, Poll},
     time::SystemTime,
 };
+use strum_macros::Display;
 #[cfg(feature = "tokio-comp")]
 use tokio::task::JoinHandle;
 
@@ -106,10 +107,10 @@ use self::{
     connections_container::{ConnectionAndAddress, ConnectionType, ConnectionsMap},
     connections_logic::connect_and_check,
 };
+use crate::types::RetryMethod;
 
-pub(crate) const MUTEX_WRITE_ERR: &str = "Failed to obtain write lock for mutex. Poisoned mutex";
-pub(crate) const MUTEX_READ_ERR: &str = "Failed to obtain read lock for mutex. Poisoned mutex";
-
+pub(crate) const MUTEX_READ_ERR: &str = "Failed to obtain read lock. Poisoned mutex?";
+const MUTEX_WRITE_ERR: &str = "Failed to obtain write lock. Poisoned mutex?";
 /// This represents an async Redis Cluster connection. It stores the
 /// underlying connections maintained for each node in the cluster, as well
 /// as common parameters for connecting to nodes and executing commands.
@@ -294,8 +295,7 @@ where
             })
             .map(|response| match response {
                 Response::ClusterScanResult(new_scan_state_ref, key) => (new_scan_state_ref, key),
-                Response::Single(_) => unreachable!(),
-                Response::Multiple(_) => unreachable!(),
+                Response::Single(_) | Response::Multiple(_) => unreachable!(),
             })
     }
 
@@ -332,8 +332,7 @@ where
             })
             .map(|response| match response {
                 Response::Single(value) => value,
-                Response::Multiple(_) => unreachable!(),
-                Response::ClusterScanResult(_, _) => unreachable!(),
+                Response::ClusterScanResult(..) | Response::Multiple(_) => unreachable!(),
             })
     }
 
@@ -357,15 +356,57 @@ where
                 sender,
             })
             .await
+            .map_err(|err| {
+                RedisError::from(io::Error::new(io::ErrorKind::BrokenPipe, err.to_string()))
+            })?;
+
+        receiver
+            .await
+            .unwrap_or_else(|err| {
+                Err(RedisError::from(io::Error::new(
+                    io::ErrorKind::BrokenPipe,
+                    err.to_string(),
+                )))
+            })
+            .map(|response| match response {
+                Response::Multiple(values) => values,
+                Response::ClusterScanResult(..) | Response::Single(_) => unreachable!(),
+            })
+    }
+    /// Update the password used to authenticate with all cluster servers
+    pub async fn update_connection_password(
+        &mut self,
+        password: Option<String>,
+    ) -> RedisResult<Value> {
+        self.route_operation_request(Operation::UpdateConnectionPassword(password))
+            .await
+    }
+
+    /// Routes an operation request to the appropriate handler.
+    async fn route_operation_request(
+        &mut self,
+        operation_request: Operation,
+    ) -> RedisResult<Value> {
+        let (sender, receiver) = oneshot::channel();
+        self.0
+            .send(Message {
+                cmd: CmdArg::OperationRequest(operation_request),
+                sender,
+            })
+            .await
             .map_err(|_| RedisError::from(io::Error::from(io::ErrorKind::BrokenPipe)))?;
 
         receiver
             .await
-            .unwrap_or_else(|_| Err(RedisError::from(io::Error::from(io::ErrorKind::BrokenPipe))))
+            .unwrap_or_else(|err| {
+                Err(RedisError::from(io::Error::new(
+                    io::ErrorKind::BrokenPipe,
+                    err.to_string(),
+                )))
+            })
             .map(|response| match response {
-                Response::Multiple(values) => values,
-                Response::Single(_) => unreachable!(),
-                Response::ClusterScanResult(_, _) => unreachable!(),
+                Response::Single(values) => values,
+                Response::ClusterScanResult(..) | Response::Multiple(_) => unreachable!(),
             })
     }
 }
@@ -410,7 +451,7 @@ type ConnectionsContainer<C> =
 
 pub(crate) struct InnerCore<C> {
     pub(crate) conn_lock: StdRwLock<ConnectionsContainer<C>>,
-    cluster_params: ClusterParams,
+    cluster_params: StdRwLock<ClusterParams>,
     pending_requests: Mutex<Vec<PendingRequest<C>>>,
     slot_refresh_state: SlotRefreshState,
     initial_nodes: Vec<ConnectionInfo>,
@@ -425,12 +466,35 @@ impl<C> InnerCore<C>
 where
     C: ConnectionLike + Connect + Clone + Send + Sync + 'static,
 {
+    fn get_cluster_param<T, F>(&self, f: F) -> Result<T, RedisError>
+    where
+        F: FnOnce(&ClusterParams) -> T,
+        T: Clone,
+    {
+        self.cluster_params
+            .read()
+            .map(|guard| f(&guard).clone())
+            .map_err(|_| RedisError::from((ErrorKind::ClientError, MUTEX_READ_ERR)))
+    }
+
+    fn set_cluster_param<F>(&self, f: F) -> Result<(), RedisError>
+    where
+        F: FnOnce(&mut ClusterParams),
+    {
+        self.cluster_params
+            .write()
+            .map(|mut params| {
+                f(&mut params);
+            })
+            .map_err(|_| RedisError::from((ErrorKind::ClientError, MUTEX_WRITE_ERR)))
+    }
+
     // return address of node for slot
     pub(crate) async fn get_address_from_slot(
         &self,
         slot: u16,
         slot_addr: SlotAddr,
-    ) -> Option<String> {
+    ) -> Option<Arc<String>> {
         self.conn_lock
             .read()
             .expect(MUTEX_READ_ERR)
@@ -478,7 +542,7 @@ where
     }
 
     // return slots of node
-    pub(crate) async fn get_slots_of_address(&self, node_address: &str) -> Vec<u16> {
+    pub(crate) async fn get_slots_of_address(&self, node_address: Arc<String>) -> Vec<u16> {
         self.conn_lock
             .read()
             .expect(MUTEX_READ_ERR)
@@ -615,6 +679,14 @@ enum CmdArg<C> {
         // struct containing the arguments for the cluster scan command - scan state cursor, match pattern, count and object type.
         cluster_scan_args: ClusterScanArgs,
     },
+    // Operational requests which are connected to the internal state of the connection and not send as a command to the server.
+    OperationRequest(Operation),
+}
+
+// Operation requests which are connected to the internal state of the connection and not send as a command to the server.
+#[derive(Clone)]
+enum Operation {
+    UpdateConnectionPassword(Option<String>),
 }
 
 fn route_for_pipeline(pipeline: &crate::Pipeline) -> RedisResult<Option<Route>> {
@@ -656,16 +728,17 @@ fn route_for_pipeline(pipeline: &crate::Pipeline) -> RedisResult<Option<Route>> 
 }
 
 fn boxed_sleep(duration: Duration) -> BoxFuture<'static, ()> {
-    #[cfg(feature = "tokio-comp")]
-    return Box::pin(tokio::time::sleep(duration));
+    Box::pin(tokio::time::sleep(duration))
 }
 
+#[derive(Debug, Display)]
 pub(crate) enum Response {
     Single(Value),
     ClusterScanResult(ScanStateRC, Vec<Value>),
     Multiple(Vec<Value>),
 }
 
+#[derive(Debug)]
 pub(crate) enum OperationTarget {
     Node { address: String },
     FanOut,
@@ -679,7 +752,26 @@ impl From<String> for OperationTarget {
     }
 }
 
-struct Message<C> {
+/// Represents a node to which a `MOVED` or `ASK` error redirects.
+#[derive(Clone, Debug)]
+pub(crate) struct RedirectNode {
+    /// The address of the redirect node.
+    pub address: String,
+    /// The slot of the redirect node.
+    pub slot: u16,
+}
+
+impl RedirectNode {
+    /// Constructs a `RedirectNode` from an optional tuple containing an address and a slot number.
+    pub(crate) fn from_option_tuple(option: Option<(&str, u16)>) -> Option<Self> {
+        option.map(|(address, slot)| RedirectNode {
+            address: address.to_string(),
+            slot,
+        })
+    }
+}
+
+struct Message<C: Sized> {
     cmd: CmdArg<C>,
     sender: oneshot::Sender<RedisResult<Response>>,
 }
@@ -740,6 +832,10 @@ impl<C> RequestInfo<C> {
                 CmdArg::ClusterScan { .. } => {
                     unreachable!()
                 }
+                // Operation requests are not routed.
+                CmdArg::OperationRequest(_) => {
+                    unreachable!()
+                }
             }
         }
     }
@@ -774,6 +870,10 @@ impl<C> RequestInfo<C> {
             CmdArg::ClusterScan { .. } => {
                 unreachable!()
             }
+            // Operation requests are not routed.
+            CmdArg::OperationRequest { .. } => {
+                unreachable!()
+            }
         }
     }
 }
@@ -789,6 +889,10 @@ pin_project! {
         Sleep {
             #[pin]
             sleep: BoxFuture<'static, ()>,
+        },
+        UpdateMoved {
+            #[pin]
+            future: BoxFuture<'static, RedisResult<()>>,
         },
     }
 }
@@ -826,6 +930,7 @@ enum Next<C> {
         // if not set, then a slot refresh should happen without sending a request afterwards
         request: Option<PendingRequest<C>>,
         sleep_duration: Option<Duration>,
+        moved_redirect: Option<RedirectNode>,
     },
     ReconnectToInitialNodes {
         // if not set, then a reconnect should happen without sending a request afterwards
@@ -853,8 +958,25 @@ impl<C> Future for Request<C> {
                 }
                 .into();
             }
+            RequestStateProj::UpdateMoved { future } => {
+                if let Err(err) = ready!(future.poll(cx)) {
+                    // Updating the slot map based on the MOVED error is an optimization.
+                    // If it fails, proceed by retrying the request with the redirected node,
+                    // and allow the slot refresh task to correct the slot map.
+                    info!(
+                        "Failed to update the slot map based on the received MOVED error.
+                        Error: {err:?}"
+                    );
+                }
+                if let Some(request) = self.project().request.take() {
+                    return Next::Retry { request }.into();
+                } else {
+                    return Next::Done.into();
+                }
+            }
             _ => panic!("Request future must be Some"),
         };
+
         match ready!(future.poll(cx)) {
             Ok(item) => {
                 self.respond(Ok(item));
@@ -867,16 +989,17 @@ impl<C> Future for Request<C> {
                     let retry_method = err.retry_method();
                     let next = if err.kind() == ErrorKind::AllConnectionsUnavailable {
                         Next::ReconnectToInitialNodes { request: None }.into()
-                    } else if matches!(err.retry_method(), crate::types::RetryMethod::MovedRedirect)
+                    } else if matches!(err.retry_method(), RetryMethod::MovedRedirect)
                         || matches!(target, OperationTarget::NotFound)
                     {
                         Next::RefreshSlots {
                             request: None,
                             sleep_duration: None,
+                            moved_redirect: RedirectNode::from_option_tuple(err.redirect_node()),
                         }
                         .into()
-                    } else if matches!(retry_method, crate::types::RetryMethod::Reconnect)
-                        || matches!(retry_method, crate::types::RetryMethod::ReconnectAndRetry)
+                    } else if matches!(retry_method, RetryMethod::Reconnect)
+                        || matches!(retry_method, RetryMethod::ReconnectAndRetry)
                     {
                         if let OperationTarget::Node { address } = target {
                             Next::Reconnect {
@@ -920,6 +1043,7 @@ impl<C> Future for Request<C> {
                         return Next::RefreshSlots {
                             request: Some(request),
                             sleep_duration: Some(sleep_duration),
+                            moved_redirect: None,
                         }
                         .into();
                     }
@@ -928,7 +1052,7 @@ impl<C> Future for Request<C> {
                 warn!("Received request error {} on node {:?}.", err, address);
 
                 match err.retry_method() {
-                    crate::types::RetryMethod::AskRedirect => {
+                    RetryMethod::AskRedirect => {
                         let mut request = this.request.take().unwrap();
                         request.info.set_redirect(
                             err.redirect_node()
@@ -936,8 +1060,9 @@ impl<C> Future for Request<C> {
                         );
                         Next::Retry { request }.into()
                     }
-                    crate::types::RetryMethod::MovedRedirect => {
+                    RetryMethod::MovedRedirect => {
                         let mut request = this.request.take().unwrap();
+                        let redirect_node = err.redirect_node();
                         request.info.set_redirect(
                             err.redirect_node()
                                 .map(|(node, _slot)| Redirect::Moved(node.to_string())),
@@ -945,10 +1070,11 @@ impl<C> Future for Request<C> {
                         Next::RefreshSlots {
                             request: Some(request),
                             sleep_duration: None,
+                            moved_redirect: RedirectNode::from_option_tuple(redirect_node),
                         }
                         .into()
                     }
-                    crate::types::RetryMethod::WaitAndRetry => {
+                    RetryMethod::WaitAndRetry => {
                         let sleep_duration = this.retry_params.wait_time_for_retry(request.retry);
                         // Sleep and retry.
                         this.future.set(RequestState::Sleep {
@@ -956,34 +1082,31 @@ impl<C> Future for Request<C> {
                         });
                         self.poll(cx)
                     }
-                    crate::types::RetryMethod::Reconnect
-                    | crate::types::RetryMethod::ReconnectAndRetry => {
+                    RetryMethod::Reconnect | RetryMethod::ReconnectAndRetry => {
                         let mut request = this.request.take().unwrap();
                         // TODO should we reset the redirect here?
                         request.info.reset_routing();
                         warn!("disconnected from {:?}", address);
-                        let should_retry = matches!(
-                            err.retry_method(),
-                            crate::types::RetryMethod::ReconnectAndRetry
-                        );
+                        let should_retry =
+                            matches!(err.retry_method(), RetryMethod::ReconnectAndRetry);
                         Next::Reconnect {
                             request: should_retry.then_some(request),
                             target: address,
                         }
                         .into()
                     }
-                    crate::types::RetryMethod::WaitAndRetryOnPrimaryRedirectOnReplica => {
+                    RetryMethod::WaitAndRetryOnPrimaryRedirectOnReplica => {
                         Next::RetryBusyLoadingError {
                             request: this.request.take().unwrap(),
                             address,
                         }
                         .into()
                     }
-                    crate::types::RetryMethod::RetryImmediately => Next::Retry {
+                    RetryMethod::RetryImmediately => Next::Retry {
                         request: this.request.take().unwrap(),
                     }
                     .into(),
-                    crate::types::RetryMethod::NoRetry => {
+                    RetryMethod::NoRetry => {
                         self.respond(Err(err));
                         Next::Done.into()
                     }
@@ -1051,7 +1174,7 @@ where
                 cluster_params.read_from_replicas,
                 0,
             )),
-            cluster_params: cluster_params.clone(),
+            cluster_params: StdRwLock::new(cluster_params.clone()),
             pending_requests: Mutex::new(Vec::new()),
             slot_refresh_state: SlotRefreshState::new(slots_refresh_rate_limiter),
             initial_nodes: initial_nodes.to_vec(),
@@ -1211,10 +1334,17 @@ where
     // Being used when all cluster connections are unavailable.
     fn reconnect_to_initial_nodes(inner: Arc<InnerCore<C>>) -> impl Future<Output = ()> {
         let inner = inner.clone();
-        async move {
+        let cluster_params = match inner.get_cluster_param(|params| params.clone()) {
+            Ok(params) => params,
+            Err(err) => {
+                warn!("Failed to get cluster params: {}", err);
+                return async {}.boxed();
+            }
+        };
+        Box::pin(async move {
             let connection_map = match Self::create_initial_connections(
                 &inner.initial_nodes,
-                &inner.cluster_params,
+                &cluster_params,
                 inner.glide_connection_options.clone(),
             )
             .await
@@ -1238,7 +1368,7 @@ where
             {
                 warn!("Can't refresh slots with initial nodes: `{err}`");
             };
-        }
+        })
     }
 
     // Validate all existing user connections and try to reconnect if necessary.
@@ -1250,16 +1380,11 @@ where
         let mut all_valid_conns = HashMap::new();
         // prep connections and clean out these w/o assigned slots, as we might have established connections to unwanted hosts
         let mut nodes_to_delete = Vec::new();
-        let all_nodes_with_slots: HashSet<String>;
+        let all_nodes_with_slots: HashSet<Arc<String>>;
         {
             let connections_container = inner.conn_lock.read().expect(MUTEX_READ_ERR);
 
-            all_nodes_with_slots = connections_container
-                .slot_map
-                .addresses_for_all_nodes()
-                .iter()
-                .map(|addr| String::from(*addr))
-                .collect();
+            all_nodes_with_slots = connections_container.slot_map.all_node_addresses();
 
             connections_container
                 .all_node_connections()
@@ -1291,8 +1416,8 @@ where
         addrs_to_refresh.extend(
             all_nodes_with_slots
                 .iter()
-                .filter(|addr| !all_valid_conns.contains_key(*addr))
-                .cloned(),
+                .filter(|addr| !all_valid_conns.contains_key(addr.as_str()))
+                .map(|addr| addr.to_string()),
         );
 
         if !addrs_to_refresh.is_empty() {
@@ -1329,7 +1454,7 @@ where
                 };
 
                 // Override subscriptions for this connection
-                let mut cluster_params = inner.cluster_params.clone();
+                let mut cluster_params = inner.cluster_params.read().expect(MUTEX_READ_ERR).clone();
                 let subs_guard = inner.subscriptions_by_address.read().await;
                 cluster_params.pubsub_subscriptions = subs_guard.get(&address).cloned();
                 drop(subs_guard);
@@ -1493,7 +1618,7 @@ where
         {
             return Ok(());
         }
-        let mut skip_slots_refresh = false;
+        let mut should_refresh_slots = true;
         if *policy == RefreshPolicy::Throttable {
             // Check if the current slot refresh is triggered before the wait duration has passed
             let last_run_rlock = last_run.read().await;
@@ -1512,13 +1637,13 @@ where
                 if passed_time <= wait_duration {
                     debug!("Skipping slot refresh as the wait duration hasn't yet passed. Passed time = {:?},
                             Wait duration = {:?}", passed_time, wait_duration);
-                    skip_slots_refresh = true;
+                    should_refresh_slots = false;
                 }
             }
         }
 
         let mut res = Ok(());
-        if !skip_slots_refresh {
+        if should_refresh_slots {
             let retry_strategy = ExponentialFactorBackoff::from_millis(
                 DEFAULT_REFRESH_SLOTS_RETRY_BASE_DURATION_MILLIS,
                 DEFAULT_REFRESH_SLOTS_RETRY_BASE_FACTOR,
@@ -1547,28 +1672,46 @@ where
         res
     }
 
+    /// Determines if the cluster topology has changed and refreshes slots and subscriptions if needed.
+    /// Returns `RedisResult` with `true` if changes were detected and slots were refreshed,
+    /// or `false` if no changes were found. Raises an error if refreshing the topology fails.
     pub(crate) async fn check_topology_and_refresh_if_diff(
         inner: Arc<InnerCore<C>>,
         policy: &RefreshPolicy,
-    ) -> bool {
+    ) -> RedisResult<bool> {
         let topology_changed = Self::check_for_topology_diff(inner.clone()).await;
         if topology_changed {
-            let _ = Self::refresh_slots_and_subscriptions_with_retries(inner.clone(), policy).await;
+            Self::refresh_slots_and_subscriptions_with_retries(inner.clone(), policy).await?;
         }
-        topology_changed
+        Ok(topology_changed)
     }
 
     async fn periodic_topology_check(inner: Arc<InnerCore<C>>, interval_duration: Duration) {
         loop {
             let _ = boxed_sleep(interval_duration).await;
-            let topology_changed =
-                Self::check_topology_and_refresh_if_diff(inner.clone(), &RefreshPolicy::Throttable)
-                    .await;
-            if !topology_changed {
-                // This serves as a safety measure for validating pubsub subscriptions state in case it has drifted
-                // while topology stayed the same.
-                // For example, a failed attempt to refresh a connection which is triggered from refresh_pubsub_subscriptions(),
-                // might leave a node unconnected indefinitely in case topology is stable and no request are attempted to this node.
+            // Check and refresh topology if needed
+            let should_refresh_pubsub = match Self::check_topology_and_refresh_if_diff(
+                inner.clone(),
+                &RefreshPolicy::Throttable,
+            )
+            .await
+            {
+                Ok(topology_changed) => !topology_changed,
+                Err(err) => {
+                    warn!(
+                        "Failed to refresh slots during periodic topology checks:\n{:?}",
+                        err
+                    );
+                    true
+                }
+            };
+
+            // Refresh pubsub subscriptions if topology wasn't changed or an error occurred.
+            // This serves as a safety measure for validating pubsub subscriptions state in case it has drifted
+            // while topology stayed the same.
+            // For example, a failed attempt to refresh a connection which is triggered from refresh_pubsub_subscriptions(),
+            // might leave a node unconnected indefinitely in case topology is stable and no request are attempted to this node.
+            if should_refresh_pubsub {
                 Self::refresh_pubsub_subscriptions(inner.clone()).await;
             }
         }
@@ -1591,7 +1734,9 @@ where
     }
 
     async fn refresh_pubsub_subscriptions(inner: Arc<InnerCore<C>>) {
-        if inner.cluster_params.protocol != crate::types::ProtocolVersion::RESP3 {
+        if inner.cluster_params.read().expect(MUTEX_READ_ERR).protocol
+            != crate::types::ProtocolVersion::RESP3
+        {
             return;
         }
 
@@ -1679,10 +1824,7 @@ where
     /// Returns true if change was detected, otherwise false.
     async fn check_for_topology_diff(inner: Arc<InnerCore<C>>) -> bool {
         let num_of_nodes = inner.conn_lock.read().expect(MUTEX_READ_ERR).len();
-        // TODO: Starting from Rust V1.67, integers has logarithms support.
-        // When we no longer need to support Rust versions < 1.67, remove fast_math and transition to the ilog2 function.
-        let num_of_nodes_to_query =
-            std::cmp::max(fast_math::log2_raw(num_of_nodes as f32) as usize, 1);
+        let num_of_nodes_to_query = std::cmp::max(num_of_nodes.ilog2() as usize, 1);
         let (res, failed_connections) = calculate_topology_from_random_nodes(
             &inner,
             num_of_nodes_to_query,
@@ -1742,9 +1884,7 @@ where
                 .await
                 .0?;
         // Create a new connection vector of the found nodes
-        let mut nodes = new_slots.values().flatten().collect::<Vec<_>>();
-        nodes.sort_unstable();
-        nodes.dedup();
+        let nodes = new_slots.all_node_addresses();
         let nodes_len = nodes.len();
         let addresses_and_connections_iter = stream::iter(nodes)
             .fold(
@@ -1752,6 +1892,7 @@ where
                 |mut addrs_and_conns, addr| {
                     let inner = inner.clone();
                     async move {
+                        let addr = addr.to_string();
                         if let Some(node) = inner
                             .conn_lock
                             .read()
@@ -1763,7 +1904,7 @@ where
                         }
                         // If it's a DNS endpoint, it could have been stored in the existing connections vector using the resolved IP address instead of the DNS endpoint's name.
                         // We shall check if a connection is already exists under the resolved IP name.
-                        let Some((host, port)) = get_host_and_port_from_addr(addr) else {
+                        let Some((host, port)) = get_host_and_port_from_addr(&addr) else {
                             addrs_and_conns.push((addr, None));
                             return addrs_and_conns;
                         };
@@ -1790,12 +1931,14 @@ where
             .fold(
                 ConnectionsMap(DashMap::with_capacity(nodes_len)),
                 |connections, (addr, node)| async {
-                    let mut cluster_params = inner.cluster_params.clone();
+                    let mut cluster_params = inner
+                        .get_cluster_param(|params| params.clone())
+                        .expect(MUTEX_READ_ERR);
                     let subs_guard = inner.subscriptions_by_address.read().await;
-                    cluster_params.pubsub_subscriptions = subs_guard.get(addr).cloned();
+                    cluster_params.pubsub_subscriptions = subs_guard.get(&addr).cloned();
                     drop(subs_guard);
                     let node = get_or_create_conn(
-                        addr,
+                        &addr,
                         node,
                         &cluster_params,
                         RefreshConnectionType::AllConnections,
@@ -1803,7 +1946,7 @@ where
                     )
                     .await;
                     if let Ok(node) = node {
-                        connections.0.insert(addr.into(), node);
+                        connections.0.insert(addr, node);
                     }
                     connections
                 },
@@ -1811,15 +1954,94 @@ where
             .await;
 
         info!("refresh_slots found nodes:\n{new_connections}");
-        // Replace the current slot map and connection vector with the new ones
+        // Reset the current slot map and connection vector with the new ones
         let mut write_guard = inner.conn_lock.write().expect(MUTEX_WRITE_ERR);
+        let read_from_replicas = inner
+            .get_cluster_param(|params| params.read_from_replicas)
+            .expect(MUTEX_READ_ERR);
         *write_guard = ConnectionsContainer::new(
             new_slots,
             new_connections,
-            inner.cluster_params.read_from_replicas,
+            read_from_replicas,
             topology_hash,
         );
         Ok(())
+    }
+
+    /// Handles MOVED errors by updating the client's slot and node mappings based on the new primary's role:
+    ///
+    /// 1. **No Change**: If the new primary is already the current slot owner, no updates are needed.
+    /// 2. **Failover**: If the new primary is a replica within the same shard (indicating a failover),
+    ///    the slot ownership is updated by promoting the replica to the primary in the existing shard addresses.
+    /// 3. **Slot Migration**: If the new primary is an existing primary in another shard, this indicates a slot migration,
+    ///    and the slot mapping is updated to point to the new shard addresses.
+    /// 4. **Replica Moved to a Different Shard**: If the new primary is a replica in a different shard, it can be due to:
+    ///    - The replica became the primary of its shard after a failover, with new slots migrated to it.
+    ///    - The replica has moved to a different shard as the primary.
+    ///      Since further information is unknown, the replica is removed from its original shard and added as the primary of a new shard.
+    /// 5. **New Node**: If the new primary is unknown, it is added as a new node in a new shard, possibly indicating scale-out.
+    ///
+    /// # Arguments
+    /// * `inner` - Shared reference to InnerCore containing connection and slot state.
+    /// * `slot` - The slot number reported as moved.
+    /// * `new_primary` - The address of the node now responsible for the slot.
+    ///
+    /// # Returns
+    /// * `RedisResult<()>` indicating success or failure in updating slot mappings.
+    async fn update_upon_moved_error(
+        inner: Arc<InnerCore<C>>,
+        slot: u16,
+        new_primary: Arc<String>,
+    ) -> RedisResult<()> {
+        let curr_shard_addrs = inner
+            .conn_lock
+            .read()
+            .expect(MUTEX_READ_ERR)
+            .slot_map
+            .shard_addrs_for_slot(slot);
+        // let curr_shard_addrs = connections_container.slot_map.shard_addrs_for_slot(slot);
+        // Check if the new primary is part of the current shard and update if required
+        if let Some(curr_shard_addrs) = curr_shard_addrs {
+            match curr_shard_addrs.attempt_shard_role_update(new_primary.clone()) {
+                // Scenario 1: No changes needed as the new primary is already the current slot owner.
+                // Scenario 2: Failover occurred and the new primary was promoted from a replica.
+                ShardUpdateResult::AlreadyPrimary | ShardUpdateResult::Promoted => return Ok(()),
+                // The node was not found in this shard, proceed with further scenarios.
+                ShardUpdateResult::NodeNotFound => {}
+            }
+        }
+
+        // Scenario 3 & 4: Check if the new primary exists in other shards
+
+        let mut wlock_conn_container = inner.conn_lock.write().expect(MUTEX_READ_ERR);
+        let mut nodes_iter = wlock_conn_container.slot_map_nodes();
+        for (node_addr, shard_addrs_arc) in &mut nodes_iter {
+            if node_addr == new_primary {
+                let is_existing_primary = shard_addrs_arc.primary().eq(&new_primary);
+                if is_existing_primary {
+                    // Scenario 3: Slot Migration - The new primary is an existing primary in another shard
+                    // Update the associated addresses for `slot` to `shard_addrs`.
+                    drop(nodes_iter);
+                    return wlock_conn_container
+                        .slot_map
+                        .update_slot_range(slot, shard_addrs_arc.clone());
+                } else {
+                    // Scenario 4: The MOVED error redirects to `new_primary` which is known as a replica in a shard that doesn’t own `slot`.
+                    // Remove the replica from its existing shard and treat it as a new node in a new shard.
+                    shard_addrs_arc.remove_replica(new_primary.clone())?;
+                    drop(nodes_iter);
+                    return wlock_conn_container
+                        .slot_map
+                        .add_new_primary(slot, new_primary);
+                }
+            }
+        }
+
+        drop(nodes_iter);
+        // Scenario 5: New Node - The new primary is not present in the current slots map, add it as a primary of a new shard.
+        wlock_conn_container
+            .slot_map
+            .add_new_primary(slot, new_primary)
     }
 
     async fn execute_on_multiple_nodes<'a>(
@@ -2005,6 +2227,13 @@ where
                     Err(err) => Err((OperationTarget::FanOut, err)),
                 }
             }
+            CmdArg::OperationRequest(operation_request) => match operation_request {
+                Operation::UpdateConnectionPassword(password) => {
+                    core.set_cluster_param(|params| params.password = password)
+                        .expect(MUTEX_WRITE_ERR);
+                    Ok(Response::Single(Value::Okay))
+                }
+            },
         }
     }
 
@@ -2099,7 +2328,7 @@ where
         let (address, mut conn) = match conn_check {
             ConnectionCheck::Found((address, connection)) => (address, connection.await),
             ConnectionCheck::OnlyAddress(addr) => {
-                let mut this_conn_params = core.cluster_params.clone();
+                let mut this_conn_params = core.get_cluster_param(|params| params.clone())?;
                 let subs_guard = core.subscriptions_by_address.read().await;
                 this_conn_params.pubsub_subscriptions = subs_guard.get(addr.as_str()).cloned();
                 drop(subs_guard);
@@ -2197,6 +2426,7 @@ where
         info: RequestInfo<C>,
         address: String,
         retry: u32,
+        retry_params: RetryParams,
     ) -> OperationResult {
         let is_primary = core
             .conn_lock
@@ -2213,7 +2443,7 @@ where
                 .remove_node(&address);
         } else {
             // If the connection is primary, just sleep and retry
-            let sleep_duration = core.cluster_params.retry_params.wait_time_for_retry(retry);
+            let sleep_duration = retry_params.wait_time_for_retry(retry);
             boxed_sleep(sleep_duration).await;
         }
 
@@ -2221,8 +2451,11 @@ where
     }
 
     fn poll_complete(&mut self, cx: &mut task::Context<'_>) -> Poll<PollFlushAction> {
+        let retry_params = self
+            .inner
+            .get_cluster_param(|params| params.retry_params.clone())
+            .expect(MUTEX_READ_ERR);
         let mut poll_flush_action = PollFlushAction::None;
-
         let mut pending_requests_guard = self.inner.pending_requests.lock().unwrap();
         if !pending_requests_guard.is_empty() {
             let mut pending_requests = mem::take(&mut *pending_requests_guard);
@@ -2236,7 +2469,7 @@ where
 
                 let future = Self::try_request(request.info.clone(), self.inner.clone()).boxed();
                 self.in_flight_requests.push(Box::pin(Request {
-                    retry_params: self.inner.cluster_params.retry_params.clone(),
+                    retry_params: retry_params.clone(),
                     request: Some(request),
                     future: RequestState::Future { future },
                 }));
@@ -2246,6 +2479,7 @@ where
         drop(pending_requests_guard);
 
         loop {
+            let retry_params = retry_params.clone();
             let result = match Pin::new(&mut self.in_flight_requests).poll_next(cx) {
                 Poll::Ready(Some(result)) => result,
                 Poll::Ready(None) | Poll::Pending => break,
@@ -2255,7 +2489,7 @@ where
                 Next::Retry { request } => {
                     let future = Self::try_request(request.info.clone(), self.inner.clone());
                     self.in_flight_requests.push(Box::pin(Request {
-                        retry_params: self.inner.cluster_params.retry_params.clone(),
+                        retry_params: retry_params.clone(),
                         request: Some(request),
                         future: RequestState::Future {
                             future: Box::pin(future),
@@ -2269,9 +2503,10 @@ where
                         request.info.clone(),
                         address,
                         request.retry,
+                        retry_params.clone(),
                     );
                     self.in_flight_requests.push(Box::pin(Request {
-                        retry_params: self.inner.cluster_params.retry_params.clone(),
+                        retry_params: retry_params.clone(),
                         request: Some(request),
                         future: RequestState::Future {
                             future: Box::pin(future),
@@ -2281,26 +2516,39 @@ where
                 Next::RefreshSlots {
                     request,
                     sleep_duration,
+                    moved_redirect,
                 } => {
                     poll_flush_action =
                         poll_flush_action.change_state(PollFlushAction::RebuildSlots);
-                    if let Some(request) = request {
-                        let future: RequestState<
-                            Pin<Box<dyn Future<Output = OperationResult> + Send>>,
-                        > = match sleep_duration {
-                            Some(sleep_duration) => RequestState::Sleep {
+                    let future: Option<
+                        RequestState<Pin<Box<dyn Future<Output = OperationResult> + Send>>>,
+                    > = if let Some(moved_redirect) = moved_redirect {
+                        Some(RequestState::UpdateMoved {
+                            future: Box::pin(ClusterConnInner::update_upon_moved_error(
+                                self.inner.clone(),
+                                moved_redirect.slot,
+                                moved_redirect.address.into(),
+                            )),
+                        })
+                    } else if let Some(ref request) = request {
+                        match sleep_duration {
+                            Some(sleep_duration) => Some(RequestState::Sleep {
                                 sleep: boxed_sleep(sleep_duration),
-                            },
-                            None => RequestState::Future {
+                            }),
+                            None => Some(RequestState::Future {
                                 future: Box::pin(Self::try_request(
                                     request.info.clone(),
                                     self.inner.clone(),
                                 )),
-                            },
-                        };
+                            }),
+                        }
+                    } else {
+                        None
+                    };
+                    if let Some(future) = future {
                         self.in_flight_requests.push(Box::pin(Request {
-                            retry_params: self.inner.cluster_params.retry_params.clone(),
-                            request: Some(request),
+                            retry_params,
+                            request,
                             future,
                         }));
                     }
@@ -2524,13 +2772,20 @@ where
             .ok()
             .and_then(|value| get_host_and_port_from_addr(addr).map(|(host, _)| (host, value)))
     });
+    let tls_mode = inner
+        .get_cluster_param(|params| params.tls)
+        .expect(MUTEX_READ_ERR);
+
+    let read_from_replicas = inner
+        .get_cluster_param(|params| params.read_from_replicas)
+        .expect(MUTEX_READ_ERR);
     (
         calculate_topology(
             topology_values,
             curr_retry,
-            inner.cluster_params.tls,
+            tls_mode,
             num_of_nodes_to_query,
-            inner.cluster_params.read_from_replicas,
+            read_from_replicas,
         ),
         failed_addresses,
     )